% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalding-package.R
\docType{package}
\name{causalding-package}
\alias{causalding}
\alias{causalding-package}
\title{causalding: Worked Solutions to 'A First Course in Causal Inference'}
\description{
Estimators, simulations and figure data behind the worked solutions to the homework problems in Ding (2024) \doi{10.1201/9781003484080}. Every exported function validates its arguments, and the package also writes the data drawn by the companion documentation site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caverac/causal-inference-ding}
  \item \url{https://caverac.github.io/causal-inference-ding/}
  \item Report bugs at \url{https://github.com/caverac/causal-inference-ding/issues}
}

}
\author{
\strong{Maintainer}: Carlos Vera-Ciro \email{caverac@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Carlos Vera-Ciro \email{caverac@gmail.com} [copyright holder]
}

}
\keyword{internal}
